% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_format_for_GoodReads.R
\name{GR_df_matrix}
\alias{GR_df_matrix}
\title{GR_df_matrix}
\usage{
GR_df_matrix(num_books)
}
\arguments{
\item{num_books}{insert the number of book in your whole library}
}
\value{
empty data frame with appropriate columns name for GoodReads.
}
\description{
Function create the empty data frame matrix according to sample table rom GoodReads website.
}
\examples{
GR_df_empty <- GR_df_matrix(num_books = 551)
}
