% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Library.R
\name{Make_Library}
\alias{Make_Library}
\title{Make_Library}
\usage{
Make_Library(URL, num_pages)
}
\arguments{
\item{URL}{generated adress URL to your own library from www.lubimyczytac.pl (example -> "https://lubimyczytac.pl/ksiegozbior/9YsJPg7kuPc")}

\item{num_pages}{the numbers of the pages in your library. Add extra 2 to number of pages.}
}
\value{
data frame with all necessary information about books in library
}
\description{
This function extract the basic information of the books from the popular portal of library www.lubimyczytac.pl like: Title, Authors,ISBN,Rating, Date Read, Shelves, Review.
}
\examples{
book_df <- LibBookLC::Make_Library(URL = "https://lubimyczytac.pl/ksiegozbior/9YsJPg7kuPc", num_pages = 30)
}
