% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_format_for_GoodReads.R
\name{pre_format_GR}
\alias{pre_format_GR}
\title{pre_format_GR}
\usage{
pre_format_GR(raw_data)
}
\arguments{
\item{raw_data}{input generated from Make_library function}
}
\value{
Data frame data frame with partly formated information about books in library for portal www.goodreads.com
}
\description{
This function will partly transform the data generated from www.lubimyczytac.pl which are acquired for www.goodreads.com. Function changes the system books rating and status of the book from polish to english.
}
\examples{
pre_format_book_df <- pre_format_GR(raw_data = book_df)
}
