% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_GoodReads_format.R
\name{Transform_GoodReads_format}
\alias{Transform_GoodReads_format}
\title{Transform_GoodReads_format}
\usage{
Transform_GoodReads_format(raw_data, num_books)
}
\arguments{
\item{raw_data}{input generated from Make_library function}

\item{num_books}{insert the number of book in your whole library}
}
\value{
data frame ready to import to Good reads
}
\description{
The function will  format data from lubimyczytac.pl website ready to import to Good Reads library website
}
\examples{
all_books <- Transform_GoodReads_format(raw_data = book_df, num_books = 551)
}
